- Default locale changed to ru - Added translation to admin module layouts - Some fixes to crud generator and templates - Added validation to client requests

+++ database/migrations/2019_02_06_100948_create_clients_table.php
@@ -25,7 +25,7 @@ class CreateClientsTable extends Migration {
 			$table->string('post_code')->nullable();
 			$table->string('city')->nullable();
 			$table->string('code')->nullable();
-			$table->integer('comission')->nullable();
+			$table->float('comission')->nullable();
 			$table->text('coment', 65535)->nullable();
 			$table->string('type', 45)->nullable();
             $table->timestamps();
